Re-encode the "none" data type value as '{}' (not null)
This was previously fixed for the `None()` helper in d2d127f7 by
explicitly encoding an empty map instead of nil.

Since we decode the value from `{}` to `nil`, that means re-encoding
such `Data` was a second path to the previous defect. This now solved
via a special `TypeNone` case in the encoder and includes a unit test.

This change also removes the need for the previous fix, so the
TestData_None unit test has been updated to expect nil instead of an
empty map.

Both paths should now be a tiny bit faster since the JSON-encoding for
"none" data type is hard-coded (doesn't rely on `encoding/json`) and we
don't have to allocate empty maps.